chore: remove deferred.js
Only used in one place anyways
