--- tests/util/deferred.js
@@ -1,16 +1,0 @@
-const deferred = () => {
-    let resolve;
-    let reject;
-    
-    const p = new Promise((ok, no) => {
-        resolve = ok;
-        reject = no;
-    });
-
-    p.resolve = resolve;
-    p.reject = reject;
-
-    return p;
-};
-
-export default deferred;
